REFACTOR: removed scipy dependency and replaced with euler_to_quaternion
Was having issues using scipy. The ros-melodic-husky-bringup script
uses python-scipy that it installed through apt-get (it's in the
distribution folder).

The scipy installed is version 0.19.0 and I need at least 1.2.0 to
use transforms, so wasn't able to use scipy.

Installing through pip did not solve this issue because the one
installed with apt-get kept taking precedence.